Hide Curator implementation details
