+++ zkfacade/src/main/java/com/yahoo/vespa/curator/impl/VespaZooKeeperFactory.java
@@ -1,5 +1,5 @@
 // Copyright 2019 Oath Inc. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
-package com.yahoo.vespa.curator;
+package com.yahoo.vespa.curator.impl;
 
 import org.apache.curator.utils.ZookeeperFactory;
 import org.apache.zookeeper.Watcher;
