+++ zkfacade/src/main/java/com/yahoo/vespa/curator/impl/package-info.java
@@ -1,0 +1,8 @@
+// Copyright Verizon Media. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
+/**
+ * @author mpolden
+ */
+@ExportPackage
+package com.yahoo.vespa.curator.impl;
+
+import com.yahoo.osgi.annotation.ExportPackage;
