+++ zkfacade/src/main/java/com/yahoo/vespa/curator/package-info.java
@@ -1,6 +1,6 @@
 // Copyright 2017 Yahoo Holdings. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
 @ExportPackage
-@PublicApi
+@PublicApi // TODO: Remove on Vespa 8. Only the api package in this should be public
 package com.yahoo.vespa.curator;
 import com.yahoo.api.annotations.PublicApi;
 import com.yahoo.osgi.annotation.ExportPackage;
